refactor: ViewController 하나의 컨트롤러로 통합
