--- app/src/main/java/com/techcourse/controller/LoginViewController.java
@@ -1,29 +1,0 @@
-package com.techcourse.controller;
-
-import com.interface21.context.stereotype.Controller;
-import com.interface21.web.bind.annotation.RequestMapping;
-import com.interface21.web.bind.annotation.RequestMethod;
-import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.view.JspView;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-@Controller
-public class LoginViewController {
-
-    private static final Logger log = LoggerFactory.getLogger(LoginViewController.class);
-
-    @RequestMapping(value = "/login/view", method = RequestMethod.GET)
-    public ModelAndView showLoginView(HttpServletRequest req, HttpServletResponse res) throws Exception {
-        String result = UserSession.getUserFrom(req.getSession())
-                .map(user -> {
-                    log.info("logged in {}", user.getAccount());
-                    return "redirect:/index.jsp";
-                })
-                .orElse("/login.jsp");
-
-        return new ModelAndView(new JspView(result));
-    }
-}
